update from scroll interactions
