+++ UltimateRecyclerView/app/src/main/res/layout/custom_refresh_activity.xml
@@ -52,7 +52,8 @@
         app:recyclerviewPadding="2dp"
         app:recyclerviewEmptyView="@layout/empty_view"
         app:recyclerviewFloatingActionView="@layout/floating_view"
-        android:layout_marginTop="?attr/actionBarSize"></com.marshalchen.ultimaterecyclerview.CustomUltimateRecyclerview>
+        android:layout_marginTop="?attr/actionBarSize"/>
+
     <com.marshalchen.ultimaterecyclerview.ui.floatingactionbutton.AddFloatingActionButton
         android:layout_width="wrap_content"
         android:id="@+id/custom_urv_add_floating_button"
