+++ UltimateRecyclerView/app/src/main/res/layout/listurv.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<com.marshalchen.ultimaterecyclerview.UltimateRecyclerView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:layout_marginTop="?attr/actionBarSize"
+    app:recyclerviewClipToPadding="true"
+    app:recyclerviewEmptyView="@layout/empty_view"
+    app:recyclerviewPadding="2dp"
+    android:id="@+id/scroll"/>
