--- UltimateRecyclerView/app/src/main/res/layout/simplefragment_ultimate_recycle_extend.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<com.marshalchen.ultimaterecyclerview.UltimateRecycleObservableExtendedView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/scroll"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:scrollbars="vertical"
-    app:recyclerviewClipToPadding="true"
-    app:recyclerviewDefaultSwipeColor="@array/google_colors"
-    app:recyclerviewFloatingActionView="@layout/floating_view"
-    app:recyclerviewScrollbars="vertical" />
