+++ UltimateRecyclerView/app/src/main/res/layout/simplefragment_viewpager_tabhost.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.marshalchen.ultimaterecyclerview.observables.TouchInterceptionLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<com.marshalchen.ultimaterecyclerview.Utils.TouchInterceptionLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
@@ -28,4 +28,4 @@
         android:background="@color/primary" />
 
 
-</com.marshalchen.ultimaterecyclerview.observables.TouchInterceptionLayout>
+</com.marshalchen.ultimaterecyclerview.Utils.TouchInterceptionLayout>
