+++ UltimateRecyclerView/app/src/main/res/layout/simplefragment_viewpaper_fragment_parent.xml
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
 -->
-<com.marshalchen.ultimaterecyclerview.observables.TouchInterceptionLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<com.marshalchen.ultimaterecyclerview.Utils.TouchInterceptionLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
@@ -65,4 +65,4 @@
                 android:background="@color/primary" />
         </LinearLayout>
     </FrameLayout>
-</com.marshalchen.ultimaterecyclerview.observables.TouchInterceptionLayout>
+</com.marshalchen.ultimaterecyclerview.Utils.TouchInterceptionLayout>
