--- UltimateRecyclerView/ultimaterecyclerview/src/main/java/com/marshalchen/ultimaterecyclerview/UltimateRecycleObservableExtendedView.java
@@ -1,79 +1,0 @@
-package com.marshalchen.ultimaterecyclerview;
-
-import android.content.Context;
-import android.support.v7.widget.RecyclerView;
-import android.util.AttributeSet;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewStub;
-
-import com.marshalchen.ultimaterecyclerview.Utils.Scrollable;
-import com.marshalchen.ultimaterecyclerview.observables.ObservableRecyclerView;
-import com.marshalchen.ultimaterecyclerview.ui.VerticalSwipeRefreshLayout;
-import com.marshalchen.ultimaterecyclerview.ui.floatingactionbutton.FloatingActionButton;
-
-/**
- * just another fancy name to tell it is a working tool
- * Created by hesk on 12/6/15.
- */
-public class UltimateRecycleObservableExtendedView extends UltimateRecyclerView {
-
-
-    public UltimateRecycleObservableExtendedView(Context context, AttributeSet attrs) {
-        super(context, attrs);
-    }
-
-    public UltimateRecycleObservableExtendedView(Context context, AttributeSet attrs, int defStyleAttr) {
-        super(context, attrs, defStyleAttr);
-    }
-
-    public UltimateRecycleObservableExtendedView(Context context) {
-        super(context);
-    }
-
-    /**
-     * overriding the variable
-     */
-    public ObservableRecyclerView mRecyclerView;
-
-    protected void initViews() {
-        LayoutInflater inflater = (LayoutInflater) getContext().getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-        View view = inflater.inflate(R.layout.ultimate_recycler_view_layout_scolable, this);
-        mRecyclerView = (ObservableRecyclerView) view.findViewById(R.id.ultimate_list);
-        mSwipeRefreshLayout = (VerticalSwipeRefreshLayout) view.findViewById(R.id.swipe_refresh_layout);
-        setScrollbars();
-        mSwipeRefreshLayout.setEnabled(false);
-        if (mRecyclerView != null) {
-            mRecyclerView.setClipToPadding(mClipToPadding);
-            if (mPadding != -1.1f) {
-                mRecyclerView.setPadding(mPadding, mPadding, mPadding, mPadding);
-            } else {
-                mRecyclerView.setPadding(mPaddingLeft, mPaddingTop, mPaddingRight, mPaddingBottom);
-            }
-        }
-
-        defaultFloatingActionButton = (FloatingActionButton) view.findViewById(R.id.defaultFloatingActionButton);
-        setDefaultScrollListener();
-
-        mEmpty = (ViewStub) view.findViewById(R.id.emptyview);
-        mFloatingButtonViewStub = (ViewStub) view.findViewById(R.id.floatingActionViewStub);
-
-        mEmpty.setLayoutResource(mEmptyId);
-
-        mFloatingButtonViewStub.setLayoutResource(mFloatingButtonId);
-
-        if (mEmptyId != 0)
-            mEmptyView = mEmpty.inflate();
-        mEmpty.setVisibility(View.GONE);
-
-        if (mFloatingButtonId != 0) {
-            mFloatingButtonView = mFloatingButtonViewStub.inflate();
-            mFloatingButtonView.setVisibility(View.VISIBLE);
-        }
-
-    }
-
-    public Scrollable getScrollable() {
-        return mRecyclerView;
-    }
-}
