--- UltimateRecyclerView/ultimaterecyclerview/src/main/java/com/marshalchen/ultimaterecyclerview/observables/ObservableRecyclerView.java
@@ -1,314 +1,0 @@
-package com.marshalchen.ultimaterecyclerview.observables;
-
-import android.content.Context;
-import android.os.Parcelable;
-import android.support.v7.widget.RecyclerView;
-import android.util.SparseIntArray;
-import android.view.MotionEvent;
-import android.view.ViewGroup;
-import android.support.v7.widget.LinearLayoutManager;
-import android.util.AttributeSet;
-import android.view.View;
-
-import com.marshalchen.ultimaterecyclerview.Utils.SavedStateScrolling;
-import com.marshalchen.ultimaterecyclerview.ObservableScrollState;
-import com.marshalchen.ultimaterecyclerview.ObservableScrollViewCallbacks;
-import com.marshalchen.ultimaterecyclerview.Utils.Scrollable;
-
-/**
- * RecyclerView that its scroll position can be observed.
- * Before using this, please consider to use the RecyclerView.OnScrollListener
- * provided by the support library officially.
- *
- * @url https://github.com/jjhesk
- */
-public class ObservableRecyclerView extends RecyclerView implements Scrollable {
-
-    // Fields that should be saved onSaveInstanceState
-    private int mPrevFirstVisiblePosition;
-    private int mPrevFirstVisibleChildHeight = -1;
-    private int mPrevScrolledChildrenHeight;
-    private int mPrevScrollY;
-    private int mScrollY;
-    private SparseIntArray mChildrenHeights;
-
-    // Fields that don't need to be saved onSaveInstanceState
-    private ObservableScrollViewCallbacks mCallbacks;
-    private ObservableScrollState mScrollState;
-    private boolean mFirstScroll;
-    private boolean mDragging;
-    private boolean mIntercepted;
-    private MotionEvent mPrevMoveEvent;
-    private ViewGroup mTouchInterceptionViewGroup;
-
-    public ObservableRecyclerView(Context context) {
-        super(context);
-        init();
-    }
-
-    public ObservableRecyclerView(Context context, AttributeSet attrs) {
-        super(context, attrs);
-        init();
-    }
-
-    public ObservableRecyclerView(Context context, AttributeSet attrs, int defStyle) {
-        super(context, attrs, defStyle);
-        init();
-    }
-
-    private SavedStateScrolling parcel_information;
-
-    public SavedStateScrolling getSavedState() {
-        return parcel_information;
-    }
-
-    @Override
-    public void onRestoreInstanceState(Parcelable state) {
-        SavedStateScrolling ss = (SavedStateScrolling) state;
-        mPrevFirstVisiblePosition = ss.prevFirstVisiblePosition;
-        mPrevFirstVisibleChildHeight = ss.prevFirstVisibleChildHeight;
-        mPrevScrolledChildrenHeight = ss.prevScrolledChildrenHeight;
-        mPrevScrollY = ss.prevScrollY;
-        mScrollY = ss.scrollY;
-        mChildrenHeights = ss.childrenHeights;
-        parcel_information = ss;
-        super.onRestoreInstanceState(ss.getSuperState());
-    }
-
-    @Override
-    public Parcelable onSaveInstanceState() {
-        Parcelable superState = super.onSaveInstanceState();
-        SavedStateScrolling ss = new SavedStateScrolling(superState);
-        ss.prevFirstVisiblePosition = mPrevFirstVisiblePosition;
-        ss.prevFirstVisibleChildHeight = mPrevFirstVisibleChildHeight;
-        ss.prevScrolledChildrenHeight = mPrevScrolledChildrenHeight;
-        ss.prevScrollY = mPrevScrollY;
-        ss.scrollY = mScrollY;
-        ss.childrenHeights = mChildrenHeights;
-        parcel_information = ss;
-        return ss;
-    }
-
-    @Override
-    protected void onScrollChanged(int l, int t, int oldl, int oldt) {
-        super.onScrollChanged(l, t, oldl, oldt);
-        if (mCallbacks != null) {
-            if (getChildCount() > 0) {
-                int firstVisiblePosition = getChildLayoutPosition(getChildAt(0));
-                int lastVisiblePosition = getChildLayoutPosition(getChildAt(getChildCount() - 1));
-                for (int i = firstVisiblePosition, j = 0; i <= lastVisiblePosition; i++, j++) {
-                    if (mChildrenHeights.indexOfKey(i) < 0 || getChildAt(j).getHeight() != mChildrenHeights.get(i)) {
-                        mChildrenHeights.put(i, getChildAt(j).getHeight());
-                    }
-                }
-
-                View firstVisibleChild = getChildAt(0);
-                if (firstVisibleChild != null) {
-                    if (mPrevFirstVisiblePosition < firstVisiblePosition) {
-                        // scroll down
-                        int skippedChildrenHeight = 0;
-                        if (firstVisiblePosition - mPrevFirstVisiblePosition != 1) {
-                            for (int i = firstVisiblePosition - 1; i > mPrevFirstVisiblePosition; i--) {
-                                if (0 < mChildrenHeights.indexOfKey(i)) {
-                                    skippedChildrenHeight += mChildrenHeights.get(i);
-                                } else {
-                                    // Approximate each item's height to the first visible child.
-                                    // It may be incorrect, but without this, scrollY will be broken
-                                    // when scrolling from the bottom.
-                                    skippedChildrenHeight += firstVisibleChild.getHeight();
-                                }
-                            }
-                        }
-                        mPrevScrolledChildrenHeight += mPrevFirstVisibleChildHeight + skippedChildrenHeight;
-                        mPrevFirstVisibleChildHeight = firstVisibleChild.getHeight();
-                    } else if (firstVisiblePosition < mPrevFirstVisiblePosition) {
-                        // scroll up
-                        int skippedChildrenHeight = 0;
-                        if (mPrevFirstVisiblePosition - firstVisiblePosition != 1) {
-                            for (int i = mPrevFirstVisiblePosition - 1; i > firstVisiblePosition; i--) {
-                                if (0 < mChildrenHeights.indexOfKey(i)) {
-                                    skippedChildrenHeight += mChildrenHeights.get(i);
-                                } else {
-                                    // Approximate each item's height to the first visible child.
-                                    // It may be incorrect, but without this, scrollY will be broken
-                                    // when scrolling from the bottom.
-                                    skippedChildrenHeight += firstVisibleChild.getHeight();
-                                }
-                            }
-                        }
-                        mPrevScrolledChildrenHeight -= firstVisibleChild.getHeight() + skippedChildrenHeight;
-                        mPrevFirstVisibleChildHeight = firstVisibleChild.getHeight();
-                    } else if (firstVisiblePosition == 0) {
-                        mPrevFirstVisibleChildHeight = firstVisibleChild.getHeight();
-                        mPrevScrolledChildrenHeight = 0;
-                    }
-                    if (mPrevFirstVisibleChildHeight < 0) {
-                        mPrevFirstVisibleChildHeight = 0;
-                    }
-                    mScrollY = mPrevScrolledChildrenHeight - firstVisibleChild.getTop();
-                    mPrevFirstVisiblePosition = firstVisiblePosition;
-
-                    mCallbacks.onScrollChanged(mScrollY, mFirstScroll, mDragging);
-                    if (mFirstScroll) {
-                        mFirstScroll = false;
-                    }
-
-                    if (mPrevScrollY < mScrollY) {
-                        //down
-                        mScrollState = ObservableScrollState.UP;
-                    } else if (mScrollY < mPrevScrollY) {
-                        //up
-                        mScrollState = ObservableScrollState.DOWN;
-                    } else {
-                        mScrollState = ObservableScrollState.STOP;
-                    }
-                    mPrevScrollY = mScrollY;
-                }
-            }
-        }
-    }
-
-    @Override
-    public boolean onInterceptTouchEvent(MotionEvent ev) {
-        if (mCallbacks != null) {
-            switch (ev.getActionMasked()) {
-                case MotionEvent.ACTION_DOWN:
-                    // Whether or not motion events are consumed by children,
-                    // flag initializations which are related to ACTION_DOWN events should be executed.
-                    // Because if the ACTION_DOWN is consumed by children and only ACTION_MOVEs are
-                    // passed to parent (this view), the flags will be invalid.
-                    // Also, applications might implement initialization codes to onDownMotionEvent,
-                    // so call it here.
-                    mFirstScroll = mDragging = true;
-                    mCallbacks.onDownMotionEvent();
-                    break;
-            }
-        }
-        return super.onInterceptTouchEvent(ev);
-    }
-
-    @Override
-    public boolean onTouchEvent(MotionEvent ev) {
-        if (mCallbacks != null) {
-            switch (ev.getActionMasked()) {
-                case MotionEvent.ACTION_UP:
-                case MotionEvent.ACTION_CANCEL:
-                    mIntercepted = false;
-                    mDragging = false;
-                    mCallbacks.onUpOrCancelMotionEvent(mScrollState);
-                    break;
-                case MotionEvent.ACTION_MOVE:
-                    if (mPrevMoveEvent == null) {
-                        mPrevMoveEvent = ev;
-                    }
-                    float diffY = ev.getY() - mPrevMoveEvent.getY();
-                    mPrevMoveEvent = MotionEvent.obtainNoHistory(ev);
-                    if (getCurrentScrollY() - diffY <= 0) {
-                        // Can't scroll anymore.
-
-                        if (mIntercepted) {
-                            // Already dispatched ACTION_DOWN event to parents, so stop here.
-                            return false;
-                        }
-
-                        // Apps can set the interception target other than the direct parent.
-                        final ViewGroup parent;
-                        if (mTouchInterceptionViewGroup == null) {
-                            parent = (ViewGroup) getParent();
-                        } else {
-                            parent = mTouchInterceptionViewGroup;
-                        }
-
-                        // Get offset to parents. If the parent is not the direct parent,
-                        // we should aggregate offsets from all of the parents.
-                        float offsetX = 0;
-                        float offsetY = 0;
-                        for (View v = this; v != null && v != parent; v = (View) v.getParent()) {
-                            offsetX += v.getLeft() - v.getScrollX();
-                            offsetY += v.getTop() - v.getScrollY();
-                        }
-                        final MotionEvent event = MotionEvent.obtainNoHistory(ev);
-                        event.offsetLocation(offsetX, offsetY);
-
-                        if (parent.onInterceptTouchEvent(event)) {
-                            mIntercepted = true;
-
-                            // If the parent wants to intercept ACTION_MOVE events,
-                            // we pass ACTION_DOWN event to the parent
-                            // as if these touch events just have began now.
-                            event.setAction(MotionEvent.ACTION_DOWN);
-
-                            // Return this onTouchEvent() first and set ACTION_DOWN event for parent
-                            // to the queue, to keep events sequence.
-                            post(new Runnable() {
-                                @Override
-                                public void run() {
-                                    parent.dispatchTouchEvent(event);
-                                }
-                            });
-                            return false;
-                        }
-                        // Even when this can't be scrolled anymore,
-                        // simply returning false here may cause subView's click,
-                        // so delegate it to super.
-                        return super.onTouchEvent(ev);
-                    }
-                    break;
-            }
-        }
-        return super.onTouchEvent(ev);
-    }
-
-    @Override
-    public void setScrollViewCallbacks(ObservableScrollViewCallbacks listener) {
-        mCallbacks = listener;
-    }
-
-    @Override
-    public void setTouchInterceptionViewGroup(ViewGroup viewGroup) {
-        mTouchInterceptionViewGroup = viewGroup;
-    }
-
-    @Override
-    public void scrollVerticallyTo(int y) {
-        View firstVisibleChild = getChildAt(0);
-        if (firstVisibleChild != null) {
-            int baseHeight = firstVisibleChild.getHeight();
-            int position = y / baseHeight;
-            scrollVerticallyToPosition(position);
-        }
-    }
-
-    /**
-     * <p>Same as {@linkplain #scrollToPosition(int)} but it scrolls to the position not only make
-     * the position visible.</p>
-     * <p>It depends on {@code LayoutManager} how {@linkplain #scrollToPosition(int)} works,
-     * and currently we know that {@linkplain LinearLayoutManager#scrollToPosition(int)} just
-     * make the position visible.</p>
-     * <p>In LinearLayoutManager, scrollToPositionWithOffset() is provided for scrolling to the position.
-     * This method checks which LayoutManager is set,
-     * and handles which method should be called for scrolling.</p>
-     * <p>Other know classes (StaggeredGridLayoutManager and GridLayoutManager) are not tested.</p>
-     *
-     * @param position position to scroll
-     */
-    public void scrollVerticallyToPosition(int position) {
-        LayoutManager lm = getLayoutManager();
-
-        if (lm != null && lm instanceof LinearLayoutManager) {
-            ((LinearLayoutManager) lm).scrollToPositionWithOffset(position, 0);
-        } else {
-            scrollToPosition(position);
-        }
-    }
-
-    @Override
-    public int getCurrentScrollY() {
-        return mScrollY;
-    }
-
-    private void init() {
-        mChildrenHeights = new SparseIntArray();
-    }
-
-}
