--- UltimateRecyclerView/ultimaterecyclerview/src/main/res/layout/ultimate_extended_view_layout.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<com.marshalchen.ultimaterecyclerview.UltimateRecycleObservableExtendedView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/scroll"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:scrollbars="vertical" />
