+++ UltimateRecyclerView/ultimaterecyclerview/src/main/res/layout/ultimate_recycler_view_layout.xml
@@ -13,7 +13,7 @@
             android:id="@+id/ultimate_list"
             android:focusableInTouchMode="true"
             android:layout_width="match_parent"
-            android:layout_height="match_parent"></android.support.v7.widget.RecyclerView>
+            android:layout_height="match_parent"/>
 
     </com.marshalchen.ultimaterecyclerview.ui.VerticalSwipeRefreshLayout>
