--- UltimateRecyclerView/ultimaterecyclerview/src/main/res/layout/ultimate_recycler_view_layout_scolable.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:fab="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <com.marshalchen.ultimaterecyclerview.ui.VerticalSwipeRefreshLayout
-        android:id="@+id/swipe_refresh_layout"
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent">
-
-        <com.marshalchen.ultimaterecyclerview.observables.ObservableRecyclerView
-            android:id="@+id/ultimate_list"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:focusableInTouchMode="true" />
-
-    </com.marshalchen.ultimaterecyclerview.ui.VerticalSwipeRefreshLayout>
-
-    <com.marshalchen.ultimaterecyclerview.ui.floatingactionbutton.AddFloatingActionButton
-        android:id="@+id/defaultFloatingActionButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentRight="true"
-        android:layout_marginBottom="15dp"
-        android:layout_marginRight="15dp"
-        android:visibility="gone"
-        fab:colorNormal="#ffffff"
-        fab:plusIconColor="#808080" />
-
-    <ViewStub
-        android:id="@+id/floatingActionViewStub"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentBottom="true"
-        android:layout_alignParentRight="true"
-        android:layout_marginBottom="15dp"
-        android:layout_marginRight="15dp"
-        android:visibility="gone" />
-
-    <ViewStub
-        android:id="@+id/emptyview"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_centerInParent="true"
-        android:visibility="gone" />
-</RelativeLayout>
